refactor: migrate attendance record to domain service (resolves #2989)
- extracted business logic from attendance/record/route.js into AttendanceService
- added institute boundary validations into service layer
- removed duplicate DB logic from route handler
