+++ app/api/attendance/record/route.js
@@ -1,16 +1,12 @@
 import { jsonError, jsonSuccess } from "@/lib/api-response";
-import { withErrorHandler, parseJSON } from "@/lib/error-handler";
+import { withErrorHandler } from "@/lib/error-handler";
 import { requireAuth } from "@/lib/rbac";
-import { initFirebaseAdmin, getUserProfile } from "@/lib/firebase-admin";
-import { getFirestore, FieldValue } from "firebase-admin/firestore";
-import { awardXp } from "@/lib/gamification-service";
 import { getLocalDateKey } from "@/lib/dateUtils";
 import { checkRateLimit } from "@/lib/rateLimit";
 import { AppError } from "@/lib/errors";
-import { executeSaga } from "@/lib/transactionCoordinator";
-import { connectDb } from "@/lib/mongodb";
 import { recordAttendanceSchema } from "@/lib/validations/attendance";
 import { validateRequest } from "@/lib/validations/validateRequest";
+import { AttendanceService } from "@/lib/services/attendanceService";
 
 
 export const POST = withErrorHandler(async (request) => {
@@ -46,109 +42,14 @@ export const POST = withErrorHandler(async (request) => {
   // Normalize confidence score to 0-1 range for consistency across the DB and dashboards
   const normalizedConfidence = parsedConfidence / 100;
 
-  // 4. Write attendance to Firestore (single source of truth).
-  // Use a deterministic doc id and a transaction to prevent duplicates and match client duplicate checks.
-  initFirebaseAdmin();
-  const db = getFirestore();
-
-
-  // Authoritatively fetch target student profile or use caller profile
-  const targetUid = userId || decodedToken.uid;
-  const userProfile = await getUserProfile(targetUid);
-  const callerProfile = decodedToken.uid !== targetUid ? await getUserProfile(decodedToken.uid) : userProfile;
-  const instituteId = userProfile?.instituteId || callerProfile?.instituteId || null;
-
-  // Institute boundary validation: teachers/admins can only submit attendance
-  // for students within their own institute
-  if (decodedToken.uid !== targetUid && isTeacherOrAdmin) {
-    if (!userProfile) {
-      return jsonError("Target user not found", 404);
-    }
-    if (userProfile.instituteId !== callerProfile?.instituteId) {
-      return jsonError("Forbidden: Cannot submit attendance for users outside your institute", 403);
-    }
-    if (userProfile.role !== "student") {
-      return jsonError("Forbidden: Attendance can only be submitted for students", 403);
-    }
-  }
-
-  // Use authoritative, verified data from profile to prevent client-supplied parameter spoofing
-  const resolvedName = userProfile?.fullName || (decodedToken.uid === targetUid ? (decodedToken.name || decodedToken.displayName) : null) || studentName || "Unknown User";
-  const resolvedEmail = userProfile?.email || (decodedToken.uid === targetUid ? decodedToken.email : null) || email || "unknown@learnova.edu";
-
-  const sagaResult = await executeSaga({
-    operationType: "attendance_record",
-    uid: decodedToken.uid,
-    steps: [
-      {
-        name: "write_attendance",
-        execute: async (ctx) => {
-          const docRef = db.collection("attendance_records").doc(`${userId}_${normalizedDate}`);
-          await db.runTransaction(async (transaction) => {
-            const existingDoc = await transaction.get(docRef);
-            if (existingDoc.exists) {
-              // Mark as already recorded — don't throw (idempotent)
-              ctx._alreadyRecorded = true;
-              return;
-            }
-
-            transaction.set(
-              docRef,
-              {
-                userId,
-                studentName: resolvedName,
-                email: resolvedEmail,
-                instituteId,
-                timestamp: FieldValue.serverTimestamp(),
-                date: normalizedDate,
-                status: "present",
-                confidenceScore: normalizedConfidence,
-                offlineSynced: false,
-              },
-              { merge: true },
-            );
-          });
-        },
-        compensate: null, // Attendance writes are append-only
-      },
-      {
-        name: "write_mongodb_attendance",
-        execute: async () => {
-          const mongoDB = await connectDb();
-          await mongoDB.collection("attendance").updateOne(
-            { userId, date: normalizedDate },
-            {
-              $set: {
-                userId,
-                studentName: resolvedName,
-                email: resolvedEmail,
-                instituteId,
-                timestamp: new Date(),
-                date: normalizedDate,
-                status: "present",
-                confidenceScore: normalizedConfidence,
-                offlineSynced: false,
-              },
-            },
-            { upsert: true }
-          );
-        },
-        compensate: async () => {
-          const mongoDB = await connectDb();
-          await mongoDB.collection("attendance").deleteOne({ userId, date: normalizedDate });
-        },
-      },
-      {
-        name: "award_xp",
-        execute: async () => {
-          await awardXp(userId, "attendance_marked", {
-            attendanceHour: new Date().getHours(),
-          });
-        },
-        compensate: null, // XP side-effect; failure doesn't block attendance
-      },
-    ],
-  });
+  // 4. Record attendance using the domain service
+  const sagaResult = await AttendanceService.recordAttendance({
+    userId,
+    studentName,
+    email,
+    confidenceScore,
+    normalizedDate
+  }, decodedToken);
 
   if (sagaResult.context._alreadyRecorded) {
     return jsonSuccess({ alreadyRecorded: true }, 200);
